Conflicts:
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/basic/DvBoolean.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/basic/DvState.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/DvAmount.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/DvOrdinal.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/DvProportion.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/DvQuantified.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/DvQuantity.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/datetime/DvDateTime.java
	openehr-rm-core/src/main/java/org/openehr/rm/datatypes/quantity/datetime/DvDuration.java

+++ openehr-rm-core/src/test/java/org/openehr/rm/common/archetyped/PathBasedValueSettingTest.java
@@ -152,7 +152,7 @@ public class PathBasedValueSettingTest extends TestCase {
 		list.set("/items[at0002]/value", ordinal);
 		
 		DvOrdinal actual = (DvOrdinal) ((Element) list.getItems().get(0)).getValue();
-		assertEquals("failed to set ordinal.value on list", 2, actual.getValue());
+		assertEquals("failed to set ordinal.value on list", 2, actual.getValue().intValue());
 		assertEquals("failed to set ordinal.symbol.value on list", "nope", 
 				actual.getSymbol().getValue());
 	}
